ggseg.extra: Create Brain Atlases for the 'ggsegverse' Plotting Ecosystem

Create brain atlas data sets compatible with the 'ggsegverse' plotting packages. Provides pipelines for building cortical, subcortical, white-matter tract, and cerebellar atlases from 'FreeSurfer' annotation files, 'GIFTI' and 'CIFTI' surface formats, 'neuromaps', cerebellar flatmaps, and volumetric 'NIfTI' images.

See also

Author

Maintainer: Athanasia Mo Mowinckel a.m.mowinckel@psykologi.uio.no (ORCID) (drmowinckels)

Other contributors:

Didac Vidal-Piñeiro d.v.pineiro@psykologi.uio.no (ORCID) [contributor]

John Muschelli muschellij2@gmail.com (ORCID) [contributor]

Center for Lifespan Changes in Brain and Cognition (LCBC), University of Oslo [copyright holder]